requirements.txt
crewai>=0.0.1
PyMuPDF>=1.22.0
pypdf>=3.12.0
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
tqdm>=4.0.0
pytest>=7.0.0
requests>=2.31.0
